There's some kind of error or memory leak in emily_driver. It causes the system to fail (stop transmitting debug messages) after a short amount of time when the PWM writing and XBee communications are enabled. May be related to my use of global status object. Simultaneous memory access??

+++ arduino/libraries/emilyStatus/emilyStatus.cpp
@@ -19,18 +19,18 @@ gpsData::gpsData(){
   lon_home = GPS_HOME_LON_DEFAULT;
 }
 
-void gpsData::set(double lati,double loni,double ti){
+void gpsData::set(double lati,double loni,float ti){
   set( int32_t(lati*1e7),int32_t(loni*1e7), ti);
 }
 
-void gpsData::set(int32_t lati, int32_t loni, double ti, double vi, double hdgi){
+void gpsData::set(int32_t lati, int32_t loni, float ti, float vi, float hdgi){
   v = (vi*KNOTS2MS);
   hdg = (hdgi*DEG2RAD);
   // call the base function
   set(lati,loni,ti);
 }
 
-void gpsData::set(int32_t lati,int32_t loni,double ti){
+void gpsData::set(int32_t lati,int32_t loni,float ti){
   init = true;
   new_value = 1;
   lat = lati;
